refactor: separate GasFeeUsd component
